Don't check apollo actions for FSA compliance
Apollo’s redux actions don’t conform to the flux-standard-action spec,
but zeal-redux-utils’ `createReducer` throws an error if any
non-FSA-compliant actions are dispatched.

zeal-redux-utils 0.3.1 now allows a whitelist function to be passed
which will bypass the FSA check on whitelisted actions.

We add a local `createReducer` function that supplies an appropriate
whitelist function for Apollo actions and the use that everywhere.

NOTE: This new function makes use of Ramda, so we now include that in
this project.