Use an absolute path as the eslint resolve root
Atom runs eslint from the directory containing the file being linted.

Because we were using a relative path to the `client` directory as our
resolve root, eslint was reporting errors and warnings because it
wasn’t able to resolve some imports.

A work-around would be to add an `.eslintignore` file to the root of
the project, but a better solution is to use an absolute path to the
`client` directory.

To do that, we need to use `path.resolve`, so our config file must be
JavaScript instead of JSON.