Fall back to `npm install` if yarn isn't available
Based on an idea from https://github.com/imperodesign/generator-impero/blob/development/generators/app/index.js#L329, use the `command-exists` package to check if yarn is available on the target system.  If so, install dependencies with yarn; otherwise, install them with npm.

There is a chance that yeoman's generator [will eventually support this automatically](https://github.com/yeoman/generator/issues/991).  If that happens, we can simplify what we're doing here.

Note that I check for the longer `yarnpkg` name, as there is another tool named `yarn` that may be installed on the target system, so `yarnpkg` is safer in this context.

I also removed the `npm install` instructions from the README, since that step is now performed automatically.