Add a precommit hook to warn about unformatted files
With this change, all staged *.js files checked for prettier-compatible formatting and the commit will fail if any are unformatted.

Ideally, we'd have prettier reformat the files for us, but there are issues if there are files with only partially-staged changes.  See https://github.com/okonet/lint-staged/issues/62 for details.

In the generator project itself, the git pre-commit hook will be installed in the parent project whenever `yarn` is run in the templates directory (which happens a lot during development).  However, lint-staged doesn't work in this context unless we add a `gitDir` setting to the lint-staged configuration.  However, we don't want that setting in the generated project.

To solve this, I changed the way we install the `package.json` file.  We read the template file, modify it in memory to remove the `gitDir` setting, and write it out to the target directory.

I also set the `name` property at the same time, so we no longer need to use an EJS template.  So now, the `package.json` file is always valid and we don't have to edit it in order to run `yarn` commands during development.  This is a much nicer workflow!