Use NODE_PATH to resolve modules
create-react-app was designed to look at `NODE_PATH` when resolving imports; we were making a bunch of changes in `@zeal/react-scripts` to do the same job, but the upcoming version of `@zeal/react-scripts` will no longer do that.

Instead, we add `NODE_PATH` to a `.env` file and that is used to resolve module imports.