Update eslint configuration for Webpack 2
The webpack options for resolver settings has changed with Webpack 2 (which is what the upcoming version of `@zeal/react-scripts` uses, so we need to change our configuration a bit.

Ideally, we'd just point at `@zeal/react-scripts`' webpack configuration file, but that doesn't work (yet).  See the comments in `app/templates/.eslintrc.js` for details.